chore(middleware/routes): Rename jwt, schema and routes
