+++ src/routes/series.routes.ts
@@ -1,7 +1,7 @@
 import Router from '@koa/router';
 import getSeries from '@business/getSeries.business';
 import upsertSeries from '@business/upsertSeries.business';
-import jwt from '@middleware/jwt';
+import jwt from '@middleware/authValidator';
 import authRenewer from '@middleware/authRenewer';
 import { DefaultState, Context } from 'koa';
 const routes = new Router<DefaultState, Context>();
